+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## In Progress
 
+- Fixed bug where invalid JSON was not printing line at all.
+- Added support for `zap.NewProduction` default lines.
 - Fixed bug where an extra new line was printed at end of stream.
 - Fixed bug where non-log line were printed twice.
 - Added formatting of Zap JSON line.